# hr_website/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "HR Website API",
    "version": "1.0.0",
    "description": "API documentation for the HR Website backend."
  },
  "components": {
    "schemas": {
...
      "Facility": {
        "type": "object",
        "required": [
          "id",
          "branch_id",
          "name",
          "code",
          "category",
          "quantity",
          "condition",
          "last_checked_at",
          "note",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "branch_id": {
            "type": "string",
            "format": "uuid",
            "description": "Chi nhánh sở hữu thiết bị (public.branches.id)."
          },
          "name": {
            "type": "string",
            "example": "Máy pha cà phê Breville"
          },
          "code": {
            "type": "string",
            "nullable": true,
            "description": "Mã tài sản / số serial trên tem thiết bị."
          },
          "category": {
            "type": "string",
            "enum": [
              "KITCHEN",
              "COLD_STORAGE",
              "FURNITURE",
              "ELECTRICAL",
              "CLEANING",
              "OTHER"
            ],
            "description": "Nhóm thiết bị (Khu pha chế, Bảo quản lạnh, ...)."
          },
          "quantity": {
            "type": "integer",
            "minimum": 1,
            "description": "Số lượng thiết bị cùng loại tại chi nhánh."
          },
          "condition": {
            "type": "string",
            "enum": [
              "GOOD",
              "FAIR",
              "MAINTENANCE",
              "BROKEN"
            ],
            "description": "Tình trạng: tốt, cần theo dõi, đang sửa chữa, hỏng (SCRUM-46 xử lý bảo trì)."
          },
          "last_checked_at": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "description": "Ngày kiểm tra gần nhất (YYYY-MM-DD)."
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "FacilityWriteRequest": {
        "type": "object",
        "required": [
          "branch_id",
          "name"
        ],
        "properties": {
          "branch_id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "code": {
            "type": "string",
            "maxLength": 60,
            "nullable": true
          },
          "category": {
            "type": "string",
            "enum": [
              "KITCHEN",
              "COLD_STORAGE",
              "FURNITURE",
              "ELECTRICAL",
              "CLEANING",
              "OTHER"
            ],
            "default": "OTHER"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9999,
            "default": 1
          },
          "condition": {
            "type": "string",
            "enum": [
              "GOOD",
              "FAIR",
              "MAINTENANCE",
              "BROKEN"
            ],
            "default": "GOOD"
          },
          "last_checked_at": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "note": {
            "type": "string",
            "maxLength": 500,
            "nullable": true
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Mô tả lỗi.",
            "example": "Trường `question` là bắt buộc và phải là chuỗi không rỗng."
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/facilities/{id}": {
      "put": {
        "summary": "Update a facility",
        "description": "Replaces the editable fields of one facility. Restricted to managers (`facility:manage`, SCRUM-59) and enforced by RLS; a hidden or unknown id answers 404. SCRUM-45.",
        "tags": [
          "Facilities"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FacilityWriteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated facility.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Facility"
                }
              }
            }
          },
          "400": {
            "description": "The id is not a UUID, or the request body is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The caller's role may not manage facilities (SCRUM-59), or they do not head that branch (SCRUM-61).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The facility does not exist or is not updatable by the caller.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a facility",
        "description": "Removes one facility from the registry. Restricted to managers (`facility:manage`, SCRUM-59) and enforced by RLS. SCRUM-45.",
        "tags": [
          "Facilities"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The facility was deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The id is not a UUID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The caller's role may not manage facilities (SCRUM-59), or they do not head that branch (SCRUM-61).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The facility does not exist or is not deletable by the caller.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}